feat: add inputboxs
