+++ extensions/iceworks-app/src/commands/stopCommand.ts
