+++ extensions/iceworks-app/src/quickPicks/showExtensionsQuickPick.ts
